refactor: remove note-waver
