+++ services/fog/go.mod
@@ -1,0 +1,3 @@
+module fog
+
+go 1.22.5
